requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8